double => decimal voor bedragen

+++ De Bank/Models/Account.cs
@@ -10,8 +10,8 @@ namespace De_Bank.Models
         public int Id { get; set; }
         public string AccountNumber { get; set; }
         public bool AccountLock { get; set; }
-        public double AccountBalance { get; set; }
-        public double AccountMinimum { get; set; }
+        public decimal AccountBalance { get; set; }
+        public decimal AccountMinimum { get; set; }
         public AccountType AccountType { get;set; }
         public AccountHolder AccountHolder { get; set; }
         public List<Transaction> transactions { get; set; }
